move all mars code in org.geotools to mars package
