+++ src/test/java/mars/memoryEater.java
@@ -1,4 +1,4 @@
-package org.geotools;
+package mars;
 
 import java.util.Vector;
